Address review comments.
- rename the analysis file to `XeGPUContiguityAnalysis.cpp`
- keep the file analysis only, and move all mechanism of apply
  to test pass.
- reduce comment clutters.
- replace the coalesce_hint attribute by I64Attr `contiguous_chunk`
