+++ mlir/include/mlir/Dialect/XeGPU/Transforms/Transforms.h
@@ -92,33 +92,13 @@ void populateXeGPUSgToLaneDistributeTypeConversionAndLegality(
 // Coalesce gather/scatter analysis + apply.
 //===----------------------------------------------------------------------===//
 
-/// Options controlling `runCoalesceGatherScatterAnalysis`.
-struct CoalesceGatherScatterAnalysisOptions {
-  /// Upper bound on the number of contiguous elements grouped per lane by
-  /// coalescing. Mirrors the `max-chunk-size` option of the original pass.
-  unsigned maxChunkSize = 8;
-};
-
-/// Run the AxisInfo-based coalescing analysis over `root` and stamp a
-/// `xegpu.coalesce_hint` attribute on every `xegpu.load` / `xegpu.store`
-/// the analysis classifies as coalescible. Ops with an existing non-empty
-/// `lane_data` or a non-uniform mask are skipped (no hint stamped). The hint
-/// is consumed downstream — either by `coalesceGatherScatter` or by a
-/// layout-propagation pass that reads the hint directly.
-void runCoalesceGatherScatterAnalysis(
-    Operation *root, const CoalesceGatherScatterAnalysisOptions &options = {});
-
-/// Walk `root` and turn every stamped `xegpu.coalesce_hint` into an
-/// equivalent `lane_layout` / `lane_data` / `inst_data` layout on its op,
-/// then remove the hint. A hint on an op that cannot be coalesced (e.g. one
-/// that isn't a gather/scatter) is simply dropped. Pairs with
-/// `runCoalesceGatherScatterAnalysis`.
-void coalesceGatherScatter(Operation *root);
-
-/// Walk `root` and remove any leftover `xegpu.coalesce_hint` attributes —
-/// useful as a cleanup after a propagator has decided whether to honor each
-/// hint.
-void clearCoalesceGatherScatterHints(Operation *root);
+/// Run the AxisInfo-based contiguity analysis over `root` and stamp a
+/// `contiguous_chunk` attribute on every `xegpu.load` / `xegpu.store` whose
+/// offsets are contiguous (in runs of >= 2) along the innermost dimension.
+/// The stamped value is the inner-dim contiguity; it is a target-independent
+/// property consumed downstream (e.g. to derive a `lane_data` split). Ops that
+/// already carry a `contiguous_chunk` attribute are left untouched.
+void runCoalesceGatherScatterAnalysis(Operation *root);
 
 /// Collect a set of patterns to unroll xegpu operations to a smaller shapes.
 /// Users can control whether an operation to be unrolled or not, as well as
