+++ mlir/lib/Dialect/XeGPU/Transforms/XeGPUWgToSgDistribute.cpp
@@ -833,7 +833,7 @@ struct WgToSgLoadGatherOp : public OpConversionPattern<xegpu::LoadGatherOp> {
       auto newLoadOp = xegpu::LoadGatherOp::create(
           rewriter, loc, newTy, op.getSource(), offsets, mask, chunkSizeAttr,
           op.getL1HintAttr(), op.getL2HintAttr(), op.getL3HintAttr(), newLayout,
-          /*coalesce_hint=*/nullptr);
+          /*contiguous_chunk=*/nullptr);
       newLoadOps.push_back(newLoadOp);
     }
     rewriter.replaceOpWithMultiple(op, {newLoadOps});
@@ -880,7 +880,7 @@ struct WgToSgStoreScatterOp
                                     mask, chunkSizeAttr, op.getL1HintAttr(),
                                     op.getL2HintAttr(), op.getL3HintAttr(),
                                     layout.dropSgLayoutAndData(),
-                                    /*coalesce_hint=*/nullptr);
+                                    /*contiguous_chunk=*/nullptr);
     }
     rewriter.eraseOp(op);
     return success();
