+++ mlir/test/Dialect/XeGPU/coalesce-gather-scatter-analyze.mlir
@@ -1,15 +1,15 @@
 // RUN: mlir-opt -split-input-file \
 // RUN:   -test-xegpu-coalesce-gather-scatter="analyze-only=true" %s | FileCheck %s
 
-// Analyze-only mode: stamps the `coalesce_hint` attribute on coalescible ops and
-// leaves the layout unchanged. This test pins the hint attribute contract
-// that the apply API (and downstream propagator integrations) consume.
+// Analyze-only mode: stamps the `contiguous_chunk` attribute on contiguous ops
+// and leaves the layout unchanged. This test pins the attribute contract that
+// the coalescing consumer (and downstream integrations) read.
 
 // -----
-// 1-D vector.step, fully coalescible -> hint with factor = 2 stamped.
+// 1-D vector.step, fully contiguous -> contiguous_chunk = 32 (the inner extent).
 // CHECK-LABEL: func.func @load_step_offsets(
 // CHECK: xegpu.load
-// CHECK-SAME: <{coalesce_hint = #xegpu.coalesce_hint<factor = 2 : i64>}>
+// CHECK-SAME: <{contiguous_chunk = 32 : i64}>
 // CHECK-SAME: : i64, vector<32xindex>, vector<32xi1> -> vector<32xf32>
 // CHECK-NOT: lane_data
 func.func @load_step_offsets(%ptr: i64) -> vector<32xf32> {
@@ -21,10 +21,10 @@ func.func @load_step_offsets(%ptr: i64) -> vector<32xf32> {
 }
 
 // -----
-// 2-D leading-1 dim: hint stamped on the load with factor = 2.
+// 2-D leading-1 dim: contiguity computed on the inner dim -> contiguous_chunk = 32.
 // CHECK-LABEL: func.func @load_2d_leading_unit(
 // CHECK: xegpu.load
-// CHECK-SAME: <{coalesce_hint = #xegpu.coalesce_hint<factor = 2 : i64>}>
+// CHECK-SAME: <{contiguous_chunk = 32 : i64}>
 // CHECK-SAME: : i64, vector<1x32xindex>, vector<1x32xi1> -> vector<1x32xf32>
 // CHECK-NOT: lane_data
 func.func @load_2d_leading_unit(%ptr: i64) -> vector<1x32xf32> {
@@ -37,12 +37,12 @@ func.func @load_2d_leading_unit(%ptr: i64) -> vector<1x32xf32> {
 }
 
 // -----
-// Stride-4 offsets: not coalescible, no hint stamped.
-// CHECK-LABEL: func.func @load_stride4_no_hint(
+// Stride-4 offsets: not contiguous, no attribute stamped.
+// CHECK-LABEL: func.func @load_stride4_no_chunk(
 // CHECK: xegpu.load
-// CHECK-NOT: coalesce_hint
+// CHECK-NOT: contiguous_chunk
 // CHECK-SAME: : i64, vector<32xindex>, vector<32xi1> -> vector<32xf32>
-func.func @load_stride4_no_hint(%ptr: i64) -> vector<32xf32> {
+func.func @load_stride4_no_chunk(%ptr: i64) -> vector<32xf32> {
   %c4 = arith.constant 4 : index
   %step = vector.step : vector<32xindex>
   %splat = vector.broadcast %c4 : index to vector<32xindex>
@@ -54,14 +54,13 @@ func.func @load_stride4_no_hint(%ptr: i64) -> vector<32xf32> {
 }
 
 // -----
-// All-equal offsets: classified as Broadcast by decide(); analysis returns
-// None at the stamping stage, so no hint is stamped (broadcast-load
-// rewrite was removed).
-// CHECK-LABEL: func.func @load_broadcast_offsets_no_hint(
+// All-equal offsets: inner dim is constant (not contiguous), so no attribute
+// is stamped.
+// CHECK-LABEL: func.func @load_broadcast_offsets_no_chunk(
 // CHECK: xegpu.load
-// CHECK-NOT: coalesce_hint
+// CHECK-NOT: contiguous_chunk
 // CHECK-SAME: : i64, vector<32xindex>, vector<32xi1> -> vector<32xf32>
-func.func @load_broadcast_offsets_no_hint(%ptr: i64) -> vector<32xf32> {
+func.func @load_broadcast_offsets_no_chunk(%ptr: i64) -> vector<32xf32> {
   %offsets = arith.constant dense<0> : vector<32xindex>
   %mask = arith.constant dense<true> : vector<32xi1>
   %v = xegpu.load %ptr[%offsets], %mask
@@ -70,14 +69,13 @@ func.func @load_broadcast_offsets_no_hint(%ptr: i64) -> vector<32xf32> {
 }
 
 // -----
-// Store with vector.step offsets: hint stamped with factor = 2 on the store
-// op.
-// CHECK-LABEL: func.func @store_step_hint(
+// Store with vector.step offsets: contiguous_chunk = 32 stamped on the store.
+// CHECK-LABEL: func.func @store_step_chunk(
 // CHECK: xegpu.store
-// CHECK-SAME: <{coalesce_hint = #xegpu.coalesce_hint<factor = 2 : i64>}>
+// CHECK-SAME: <{contiguous_chunk = 32 : i64}>
 // CHECK-SAME: : vector<32xf32>, i64, vector<32xindex>, vector<32xi1>
 // CHECK-NOT: lane_data
-func.func @store_step_hint(%ptr: i64, %v: vector<32xf32>) {
+func.func @store_step_chunk(%ptr: i64, %v: vector<32xf32>) {
   %offsets = vector.step : vector<32xindex>
   %mask = arith.constant dense<true> : vector<32xi1>
   xegpu.store %v, %ptr[%offsets], %mask
@@ -86,15 +84,15 @@ func.func @store_step_hint(%ptr: i64, %v: vector<32xf32>) {
 }
 
 // -----
-// A user-provided coalesce_hint takes precedence: the analysis must not
-// overwrite it, even if its own heuristic would pick a different factor.
-// CHECK-LABEL: func.func @user_hint_preserved(
+// A pre-existing contiguous_chunk takes precedence: the analysis must not
+// overwrite it.
+// CHECK-LABEL: func.func @user_chunk_preserved(
 // CHECK: xegpu.load
-// CHECK-SAME: <{coalesce_hint = #xegpu.coalesce_hint<factor = 2 : i64>}>
-func.func @user_hint_preserved(%ptr: i64) -> vector<32xf32> {
+// CHECK-SAME: <{contiguous_chunk = 2 : i64}>
+func.func @user_chunk_preserved(%ptr: i64) -> vector<32xf32> {
   %offsets = vector.step : vector<32xindex>
   %mask = arith.constant dense<true> : vector<32xi1>
-  %v = xegpu.load %ptr[%offsets], %mask <{coalesce_hint = #xegpu.coalesce_hint<factor = 2>}>
+  %v = xegpu.load %ptr[%offsets], %mask <{contiguous_chunk = 2 : i64}>
       : i64, vector<32xindex>, vector<32xi1> -> vector<32xf32>
   return %v : vector<32xf32>
 }
